# Conflicts:
#	.travis.yml
#	modules/swagger-codegen-cli/pom.xml
#	modules/swagger-codegen-maven-plugin/pom.xml
#	modules/swagger-codegen/pom.xml
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/CodegenProperty.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/DefaultGenerator.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/AbstractAdaCodegen.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/AbstractCSharpCodegen.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/AdaCodegen.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/GoClientCodegen.java
#	modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/GoServerCodegen.java
#	modules/swagger-codegen/src/test/java/io/swagger/codegen/java/JavaModelTest.java
#	modules/swagger-generator/pom.xml
#	pom.xml
#	pom.xml.circleci

+++ samples/server/petstore/rust-server/src/models.rs
@@ -392,11 +392,11 @@ impl EnumTest {
 pub struct FormatTest {
     #[serde(rename = "integer")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub integer: Option<i32>,
+    pub integer: Option<u8>,
 
     #[serde(rename = "int32")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub int32: Option<i32>,
+    pub int32: Option<u32>,
 
     #[serde(rename = "int64")]
     #[serde(skip_serializing_if="Option::is_none")]
@@ -603,7 +603,7 @@ pub struct Name {
 
     #[serde(rename = "123Number")]
     #[serde(skip_serializing_if="Option::is_none")]
-    pub _123_number: Option<i32>,
+    pub _123_number: Option<isize>,
 
 }
